Allow aliases for actors on the flag page
Instead of having actors displayed in the format:

 * user:1
 * user:2
 * user:404
 * country:uk

Aliases allow a function to be called with the flag, which could fetch
the users names from the database and transform them into:

 * User Name 1
 * User Name 2
 * user:404
 * United Kingdom

This is achieved with the following (optional) config:

    config :fun_with_flags_ui, :flag_page,
      alias_fn: [MyApp.Feature, :flag_alias]

The function expects a `%FunWithFlags.Flag{}` struct, and returns a
`%FunWithFlags.Flag{}` struct.

An example (inefficient) implementation of the
`MyApp.Feature.flag_alias/1` function could be:

    def flag_alias(flag) do
      gates = Enum.map(flag.gates, &get_gate_aliases/1)
      %{flag | gates: gates}
    end

    defp get_gate_aliases(%{for: "user:" <> id} = gate) do
      case MyApp.Repo.get(MyApp.User, id) do
        nil -> gate
        user -> %{gate | for: {"account:" <> id, user.name}}
      end
    end
    defp get_gate_aliases(%{for: "country:uk"} = gate) do
      %{gate | for: {"country:uk", "United Kingdom"}}
    end
    defp get_gate_aliases(gate), do: gate

If the `for` is a 2-tuple, in the format `{:id, :display_name}` then the
`:id` will be used for all form actions (such as enable/disable) and the
`:display_name` field will be used when actually displaying the actor
row.