Support for slash in group name
Encode slash as %2F. If group name contained a / character then
this would be interpreted by the elixir as a different routing
segment. I'm not sure if this works across all browsers or
network setups because maybe some browsers will unescape values
in the path before submitting the path or middle boxes might
rewrite the path to an unescaped form.