Make percentage input consistent
This makes the Percentage feature more consistent, such that the input is accepted as a value between 0 and 100. Previously the input was required to be a ratio (between 0 and 1), but it was then displayed as a percentage (effectively `"#{ratio / 100}%"`). With this change, the formatting of the input and later display match.